Cisco: Fix power supply monitoring
This was ignoring critical power supplies if the RPS was working.  The
point of an RPS is to keep the switch running until the power supply can
be replaced.  The comment discusses how one particular user was powering
their switches using an RPS all the time.  While that is their choice,
it is certainly unusual.  And having that NOT be an alarm breaks the
typical use case where one would want to know that a power supply is
broken.