Fix interface availability
Currently interface availability is not working correctly when
ifOperStatus is UP and ifAdminStatus is also UP.

Also, one wants to consider the number of unavailable interfaces, instead
of the number of available ones, when checking the thresholds -
otherwise we can't consistently have a reasonable default threshold.
Change this and adjust the default for the warning and critical values.